Remake configuration as overrides
Initially plan was to create configuration file if there's none, but it was reconsidered, when need for Sunshine's log path arose.

Now configuration is created using default values and configuration file only overrides values if it exists, it is not created otherwise. So there is no need for a configuration file to always exist.

Therefore, deletion event is no longer needed, and instead of it creation event is used.

Logging messages also have been edited to respect source of configuration: disk or defaults. They are now being written into file log as they are in "information" category.