Utilize Sunshine log file instead of WMI
Sunshine writes about connections to it's log. So we can check information faster and not use WMI.

WMI is not used anymore, so dependency on WmiLight is no longer needed.

FileSystemWatcher is not used, because it is not fired on log changes, because the file is not closed by Sunshine. So checking for new lines in the log file is comparable to watcher in terms of implementation, difference in performance is expected to be minimal.

In addition to previous commit, existing in project configuration file was renamed and it now serves as example.